Move notifiable_task to common and split into .h and .cpp
