+++ src/include/brain/scheduler_task.h
@@ -1,0 +1,19 @@
+//===----------------------------------------------------------------------===//
+//
+//                         Peloton
+//
+// scheduler_task.h
+//
+// Identification: src/include/brain/scheduler_task.h
+//
+// Copyright (c) 2015-2018, Carnegie Mellon University Database Group
+//
+//===----------------------------------------------------------------------===//
+
+#include "common/notifiable_task.h"
+
+namespace peloton {
+namespace brain {
+
+}
+}
