+++ src/include/network/connection_dispatcher_task.h
@@ -12,7 +12,7 @@
 
 #pragma once
 
-#include "notifiable_task.h"
+#include "common/notifiable_task.h"
 #include "network_state.h"
 #include "concurrency/epoch_manager_factory.h"
 #include "connection_handler_task.h"
